Fix holo tabs to work on phone screens.
This isn't perfect.  I think the assets need to be reworked to
have padding at the bottom to account for the bar; in the change
here the text appears on top of the bar because it is aligned
with the bottom.

Also the icons now used for tabs are smaller than the old UI,
making the spacing here strange.  For now I am dealing with all
of this by tweaking the layout dimensions.

Change-Id: I696100edba85a3f384d79aeb2538554c99d48437