stagefright aacenc/amrwbenc: Remove trailing whitespace
This was applied in AOSP master earlier, in commit
07d96f2679d47a2f7d1a8123340590e2813b3f34, but when merged
into the Google internal tree in
10781dc06cf3e482f5e5fd6425a1f9a43bef3a56, none of these
changes have been brought in. (The diff of this merge commit
and its first parent, 0d834ea, is empty.)

Therefore, reapply the rest of these changes on the files
that still are present in the current version.

Change-Id: I645a9b9134769958a11808a06163123d28d83e46