During encryption, don't throw exceptions for unknown vold state
This is a temporary fix until we get something better.  During the
encryption process, vold does not respond to commands because it is
blocked doing the encryption.  Because it doesn't respond, this code
times out, sets the state to null, and thows an exception.  For now,
just don't throw an exception if the device is encrypting.

Change-Id: I7e821ec2addd60475fb10834df533beccb2650fe